feat(tvix/eval): Add a fallible port of std's sort
Add a port of the internal sort implementation in rust's std, modified
to allow the comparator function to return a result, in preparation for
using this to implement the `sort` builtin.

Copying code over from std like this is quite unfortunate, but I
couldn't think of an alternate choice that doesn't make sacrifices in
terms of performance - this is segmented off in a corner of the
codebase, though, so if we find a better option in the future we should
ideally be able to pretty quickly replace this with that.

Change-Id: Id75316d4f1304d0311f9d2f06f42d1a13698fa29