chore(3p): obtain home-manager source from upstream directly
Instead of waiting for nixpkgs to [update home-manager], we can track
the upstream repository directly (using master as a channel seems to
be common practice) and overlay the up to date source into our nixpkgs
instance.

For //users/tazjin/home and //users/wpcarro/nixos/marcus, we need to
set `home.stateVersion`, since the default value for this option was
removed some time this summer.

[update home-manager]: https://github.com/NixOS/nixpkgs/issues/197907

Change-Id: I8c153fb7d3b55a4040652a5d619761b640119105