feat(tvix/proto/pathinfo.proto): init
This adds the PathInfo message, which hosts information about a Nix
Store path, mapping to either of a {Directory,File,Symlink}Node.

We import `castore.proto` here, so we need to have imports working.

Copy the protos into $out/tvix/proto, so the import paths relative to the
root work without importing all of depot into the store on every build.

Change-Id: I79d871b6fad450d6a4ae4101fb72c51f9a83471f