chore(3p/sources): Bump channels & overlays
The OpenSSL 3.0 security update tomorrow will require us to update to
the staging-next branch of nixpkgs, it looks like. This CL updates depot
to nixpkgs master in preparation so we can sort out any unrelated
breakage beforehand.

So far it looks to be significant changes in the grafana module which
I've disabled temporarily to be fixed in a follow-up CL, as I lack the
time to adapt our configuration currently.

Change-Id: Ie08fef49cf3db12e6b5225a8b992a990ddc5b642