refactor(tvix/eval): remove lifetime parameter from VM<'o>
Update, 2022-Nov-06:

It turns out that is possible to retain genericity over the lifetime of the observer, if that is a critical requirement.

The trick is that instead of making the VM generic, you instead make Value generic! Then you can thread any genericity you like (types, lifetimes, etc) through the entire program. And you can even thread that genericity through the rust std traits: impl<'o> PartialEq for Value<'o> { ... }.

In hindsight this is sort of obvious; as a pure functional language, the central object of Nix is its values, not its call stack.


**** original commit message below ****

This commit simply changes the type of VM::observer from

  &'o mut dyn RuntimeObserver

to

  Box<dyn RuntimeObserver>

which allows to drop the lifetime paramter from VM.

Rust's borrow checker requires a bit of trivial+local let-floating
to keep it happy; the resulting diff-noise is amplified by rustfmt
which makes the patch look considerably larger than it really is.

Calling code (such as tvix) can use Rc<RefCell<>> instead of lending
into VM::new().

Signed-off-by: Adam Joseph <adam@westernsemico.com>
Change-Id: If1e48d2c83e7e762747be4e36f1f256a1f0c04f7