feat(tvix/eval): Add proc-macros for defining builtins
Add a pair of proc macros to a new proc-macro crate,
`tvix-eval-proc-macros` - one for distributed registry of builtins and
one for collecting all those registered builtins into a single vec.

This is relatively straightforward as written, but may not
actually *work* long term, as expansion order of proc-macros is in
general unspecified - but let's give it a shot!

Change-Id: Ie4afae438914d2af93d15637151a49b4c68aa352