feat(tvix/eval): non-recursive implementation of nix_eq()
I'm not totally happy with this yet, but I don't want to hoard it in
my private tree either.  It's a necessary part of dealing with our
overflowing-the-stack situation, and in its current state it's good
enough to expose other problems, which are the ones I'm working on
right now.

Change-Id: Ib47535ba2fc77a4f1c2cc2fd23d3a879e21d8b4c
Signed-off-by: Adam Joseph <adam@westernsemico.com>