feat(tvix/eval): work around clap bug so --warnings isn't ignored
Clap tries to automatically infer the `action` from the type of the
field.  For booleans, the inferred action is `SetTrue`.  If
`default_value="true"` is included, clap will allow `--warnings
false` but interpret this as "set `warnings` to `true`".

I consider this a bug in clap.  Let's work around it so `--warnings`
is no longer ignored.

Change-Id: Iac8063fee9b8e0a32b28311d5170d4ee01fc4c25