feat(depotfmt): EXPERIMENT! lazily load deps
This is the worst possible way of using the depot-deps to format files
instead of loading everything at once.

The big biiiig disadvantage (besides the abstraction violations) is
that this will call `nix-build -A $attr` for *every* call to a lazy
tool e.g. `go fmt`. So we pay the nix eval overhead for every. single.
file. we want to format.

My proposed “solution” to this is that we’d want to provide a
different cache, one that has the binaries in a $tmp dir and iff a
binary is called for the first time, it puts a mutex, realizes,
and replaces itself. Then in the following calls it would just use the
symlink to the realized path.

Change-Id: I67b88a27e8a58aa503b9bb2403f9c994402e6dfd