feat(tvix/eval): add builtins to builtins
This is a somewhat terrifying hack that enables us to support
`builtins.builtins`, by running a "fake compilation" inside of a
suspended native thunk that can resolve the weak pointer to the
globals.

With this implementation, the thunk at `builtins.builtins` actually
resolves to the "real" `builtins` (verified with a new test).

This is kind of ugly, and it's something users shouldn't use, but
bubbling a test out of this is difficult at the moment due to a little
bit of trickery with how the spans in suspended native thunks
work (they don't), which a subsequent commit will address.

Change-Id: I67d0e93246dd5b279c960aeda00402031aa12af3