refactor(tvix/eval): non-hacky suspended native thunks
Instead of having a representation of suspended native thunks that
involves constructing a fake code chunk, make these thunks a
first-class part of the internal thunk representation.

The previous code was not that simple to understand, and actually
contained a critical bug which could lead to Tvix crashes.

This makes it possible to reproduce b/238.

Fixes: b/237.
Change-Id: I771d03864084d63953bdbb518fec94487481f839