fix(tvix/derivation): fix build after StorePath rename
This project was not previously covered by CI (fixed in this commit),
so we didn't catch breakage due to a renamed module.

This was noticed while rebasing a CL that has a dependency on this
crate in its Nix build.

There is also a test that seems to have broken (fails to compile) due
to some type changes, but I couldn't immediately see what was wrong
and commented it out. The test didn't work before either, so nothing
really changed there.

Change-Id: Ic48570b9313e5f73e14daab50cf7ea70918c94d1