fix(tvix/derivation): fix build after StorePath rename
Seems like the rename was a string replace which did not rename the
module path, which broke the build.

I didn't notice this until rebasing cl/7756 because it turns out that
this crate did not have a build target in CI, which I've now fixed by
adding a `default.nix` and renaming the crate to the correct
name (`tvix-derivation`).

Change-Id: Ic48570b9313e5f73e14daab50cf7ea70918c94d1