feat(users/Profpatsch): global shell for my userdir
For my tooling, I want to be able to use vscode language servers for
all subprojects, and the best ways to do that I’ve found so far is to
add a global shell.nix which contains the transitive closure of all
dependencies I need.

This is not /nice/ per se, but it does the job with minimal effort
right now and gives me a good development environment for all these
crazy & dumb experiments in here.

Change-Id: I717a72f490e9d58d45e4e15e9ba604c36b299814