fix(tazjin/tverskoy): disable systemd-oomd
This seems to have been turned on in ~November/December 2022, and is
wreaking havoc on my system usability as it keeps killing my X session
as soon as any kind of load happens on the system.

This can be as little as accidentally playing two YouTube videos (as
each YouTube tab can take 3-5GiB of RAM), or trying to process a
dataset locally.

I'm not sure if this is the culprit, but it sure seems like it.

Change-Id: Id742b4506262cc362c9fd3f2575aea23e5c092a7