refactor(tvix/derivation): pass string to build_store_path, not &[u8]
All of these occurences are actually sha256 digests of some fingerprint.

Instead of passing in a digest, keep the sha256 hasher in the function,
and pass in a fingerprint.

This makes it much easier to construct fingerprints using format!() in
all consumers, because we need don't need to juggle with the hasher
anymore.

Change-Id: I2dc3af2cab6cf06f55ae6cbd9a8be95faf2a07b6