refactor(tvix/eval): select multiple formals from attrs at once
When compiling formals, emit code to select all of the *required*
formals from the same set at once.

This avoids excessive cloning of the attribute set when calling a
function, which is particularly relevant for callPackage-style
function calls which often pass a large amount of non-optional
attributes from `<nixpkgs>`.

This changes no observable runtime behaviour (in particular, no
thunking/forcing differences can be observed).

This was a known optimisation, however I found out while writing it
that we can only do this for formals - not for inherits - as the thunk
order of inherits would be affected.

Change-Id: I8822714fee0b22cce500c3809b997364ccb529ed