refactor(tvix/eval): use a union representation of ops in Chunk
This will allow us to add a `data` field to the union representation
to significantly reduce the size of an `Op` in the code
chunks (eventually, from the current 16 to 1 (one!) byte).

However, this is a delicate undertaking as we can not easily verify
the validity of an OpCode and accessing these operations requires
`unsafe`, which means we're on our own. A compiler bug could lead to
segfaults as soon as data operands are used.

I believe that the cache efficiency gained by this is worth it, but it
will take several CLs to get there and prove this. It will likely take
some future pain for us to be fully confident in this.

I want to get this to work, and then tackle the known possible bugs
found along the way, and then make a decision on whether it is worth
it.

Change-Id: Ic745a5e5cf68417e79538946fbf06e2945db04a4