refactor(tvix/eval): use usize operand for OpConstant
Instead of having an operand field on this variant, write a usize into
the chunk's bytecode vector to carry the operand.

The interfaces that are growing here are subject to change, especially
in the chunk. For example, we might want to lock down the `CodeIdx`
type (return new ones in the chunk after reading, etc.), but that can
be done later.

Change-Id: Ib690dfcc4b2b0fb18714a35de2d9b2f2bf0c2292