refactor(tvix/eval): use usize operand for jumps
This removes the enum field from all jump variants in OpCode, and
instead encodes the jump arg as individual bytes in the bytecode
chunk.

This was quite tricky to get right - I actually had to draw on paper
to figure out the right offset calculation. Fortunately, our test
suite seems to cover everything that can be wrong here and I've also
added a unit test for the internal behaviour of Chunk.

Change-Id: I626f912c9a774b996ca3f9245d0f22885204bd72