refactor(tvix/eval): use usize operands to thunk/closure creation
Emitting these operations is a little more complex at first sight, as
the compiler potentially mutates them to optimise after the first
point where they are emitted.

However, on closer inspection, the operands to all three ops are the
same so they can be replaced in the compiler without consequences.

Change-Id: I7ff403fe99e72eea4bb2b73ab54c5e01e6c57cb3